+++ src/include/graphics/objects/sphere_sector.h
@@ -17,7 +17,7 @@ extern "C" {
 extern unsigned int sphere_sector_vbo;
 extern vertex3d sphere_sector_vbo_data[SPHERE_SECTOR_HEIGHT1 * SPHERE_SECTOR_WIDTH1 + SPHERE_SECTOR_WIDTH1 + 2 * SPHERE_SECTOR_HEIGHT1 + 3];
 #define SPHERE_SECTOR_IBO_DATA_SIZE (SPHERE_SECTOR_HEIGHT * SPHERE_SECTOR_WIDTH * 6 + SPHERE_SECTOR_WIDTH * 3 + 2 * SPHERE_SECTOR_HEIGHT * 3)
-extern short int sphere_sector_ibo_data[SPHERE_SECTOR_IBO_DATA_SIZE];
+extern unsigned int sphere_sector_ibo_data[SPHERE_SECTOR_IBO_DATA_SIZE];
 void init_sphere_sector_object(void);
 void free_sphere_sector_object(void);
 draw_obj make_draw_sphere_sector1f(float radius, int s, material_t _material);
